journal_search_update_commit
Updated the Journal, Magazine, & Newspaper by Title search to run a
search by keword and display the rsults of that search if the title
search does not find any matches. A fix was also added to account for a
screen scrapping issue when a descriptive display was returned in the
title search. Also, the video search for the touchkiosk was changed to
remove streaming videos.